реструктурирован файл utils.js
