+++ src/helpers/utils.js
@@ -1,6 +1,4 @@
 import dayjs from 'dayjs';
-// import duration from 'dayjs/plugin/duration';
-// import relativeTime from 'dayjs/plugin/relativeTime';
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 const updateItem = (items, update) => items.map((item) => item.id === update.id ? update : item);
@@ -21,50 +19,6 @@ const getRandomInteger = (min, max) => {
 };
 
 const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
-const getRandomStartDate = () => new Date(2024, 1, getRandomInteger(0, 31), getRandomInteger(7, 23), getRandomInteger(0, 59));
-const getRandomEndDate = () => new Date(2024, 2, getRandomInteger(0, 31), getRandomInteger(7, 23), getRandomInteger(0, 59));
-
-
-/* ---- DATES AND TIME ---- */
-function getDateTimeFullText(date) {
-  return date ? dayjs(date).format('YYYY-MM-DDTHH:mm') : '';
-}
-
-function getDateTimeShortText(date) {
-  return date ? dayjs(date).format('YYYY-MM-DD') : '';
-}
-
-function getDateTimeFieldText(date) {
-  return date ? dayjs(date).format('DD/MM/YY HH:mm') : '';
-}
-
-function getDateMonthText(date) {
-  return date ? dayjs(date).format('MMM DD') : '';
-}
-
-function getTimeText(date) {
-  return date ? dayjs(date).format('hh:mm') : '';
-}
-
-function getDuration(dateStart, dateEnd) {
-  let minutes = 0;
-  let hours = 0;
-  let days = 0;
-
-  minutes = ((new Date(dateEnd).getTime()) - (new Date(dateStart).getTime())) / (1000 * 60);
-
-  if (minutes >= 60) {
-    hours = Math.floor(minutes / 60);
-    minutes = Math.floor(minutes % 60);
-  }
-
-  if (hours > 23) {
-    days = Math.floor(hours / 24);
-    hours = Math.floor(hours % 24);
-  }
-
-  return { minutes, hours, days };
-}
 
 export {
   isEscapeKey,
@@ -73,12 +27,4 @@ export {
   sortEventsByTime,
   getRandomInteger,
   getRandomArrayElement,
-  getRandomStartDate,
-  getRandomEndDate,
-  getDateTimeFullText,
-  getDateTimeShortText,
-  getDateTimeFieldText,
-  getDateMonthText,
-  getTimeText,
-  getDuration,
 };
